pyproject.toml
[tool.poetry]
name = "documentation-assistant"
version = "0.1.0"
description = ""
authors = ["web-lizzard <czarnous.adrian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.340"
fastapi = "^0.104.1"
pydantic-settings = "^2.1.0"
openai = "^1.3.5"
chromadb = "^0.4.18"
unstructured = "^0.11.0"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.1"
pinecone-client = "^2.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
